refactor(cosmo): move all_parameters to parameter module
